Turned mechanism for fetching phonePrefixes into a custom hook
