+++ packages/lib/src/components/internal/PhoneInputNew/PhoneInput.scss
@@ -60,8 +60,8 @@
 
         .adyen-checkout-dropdown--countrycode-selector {
             border-right: 1px solid #dce0e5;
-            min-width: 130px;
-            width: 130px;
+            min-width: 134px;
+            width: 134px;
         }
 
         .adyen-checkout-input-holder--phone-input {
